chore: improve error handling
